service level test cases

+++ src/main/java/com/odschool/dtos/StudentRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class StudentRequest {
     private String studentName;
-    private float obtainMarks;
-    private float percentage;
+    private double obtainMarks;
+    private double percentage;
     private String result;
 }
