+++ src/main/java/com/odschool/dtos/StudentResponse.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class StudentResponse {
     private int id;
     private String studentName;
-    private float obtainMarks;
-    private float percentage;
+    private double obtainMarks;
+    private double percentage;
     private String result;
     private int divisionId;
 }
