+++ src/main/resources/db/migration/V4__create_subject_mark_table.sql
@@ -1,7 +1,7 @@
 create table subjectmark(
     id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     subject_name VARCHAR(50) NOT NULL,
-    marks INT NOT NULL,
+    marks DOUBLE PRECISION NOT NULL,
     student_id INT,
     CONSTRAINT fk_student
         FOREIGN KEY (student_id)
